[update_handler] Post PR review fix. Remove the GestionShopHandler and put all the relevant methods in the UpdateHandler file
