+++ config.yaml
@@ -6,11 +6,11 @@ parameters:
   shop:
     # 'helloasso', 'mangopay' and in future 'stripes', 'mollie' or other
     serviceName: 'helloasso'
-    lists:
-      ListeFormsProduit_id: 'ListeFormsProduit'
   shop_editable_config_params:
     - shop:
       - serviceName
+      - forms:
+          - products
       - mangoPay:
         - clientId
         - clientApiKey
