--- templates/update-handler.twig
@@ -1,12 +1,0 @@
-<strong>{{ _t('SHOP_UPDATE_TITLE') }}</strong>
-<br>
-<a
-        href="{{ url({handler:'gestionshop'}) }}"
-        title="{{ _t('SHOP_MANAGE') }}"
-        class="btn btn-primary btn-xs"
-        data-toggle="tooltip"
-        data-placement="right"
->
-    {{ _t('SHOP_MANAGE') }}
-</a>
-<hr />
